requirements.txt
python-dotenv
finnhub-python
avro-python3
websocket-client
confluent-kafka
confluent-kafka[avro]
requests